User role done on AutoRenew, Billing, User

+++ app/Http/Controllers/Api/AutoRenewController.php
@@ -7,7 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Models\AutoRenew;
 use Illuminate\Http\Request;
 use Validator;
-
+use App\Permissions\Permission;
 class AutoRenewController extends Controller
 {
     /**
@@ -17,8 +17,17 @@ class AutoRenewController extends Controller
      */
     public function index()
     {
-        $autoRenews = AutoRenew::all();
-        return response()->json($autoRenews);
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_AUTORENEW_INDEX)) {
+            $autoRenews = AutoRenew::all();
+            return response()->json($autoRenews);
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
     }
 
     /**
@@ -29,36 +38,45 @@ class AutoRenewController extends Controller
      */
     public function store(Request $request)
     {
-        $input = $request->all();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_AUTORENEW_STORE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'auto_renew_min_amt' => 'required|string',
-            'auto_renew_amt' => 'required|string|max:50',
-            'auto_renewal' => 'required|string|max:100',
-        ]);
-        
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
+            $validator = Validator::make($input, [
+                'auto_renew_min_amt' => 'required|string',
+                'auto_renew_amt' => 'required|string|max:50',
+                'auto_renewal' => 'required|string|max:100',
             ]);
-        }
+            
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
 
-        try {
-            $autoRenew = AutoRenew::create($input);
-            return response()->json($autoRenew);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
+            try {
+                $autoRenew = AutoRenew::create($input);
+                return response()->json($autoRenew);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "AutoRenew store error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
             }
-            else{
-                $message = "AutoRenew store error";
-            }
+        }
+        else{
             return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
             ]);
         }
     }
@@ -71,7 +89,16 @@ class AutoRenewController extends Controller
      */
     public function show(AutoRenew $autoRenew)
     {
-        return response()->json($autoRenew);
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_AUTORENEW_SHOW)) {
+            return response()->json($autoRenew);
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
     }
 
     /**
@@ -83,36 +110,45 @@ class AutoRenewController extends Controller
      */
     public function update(Request $request, AutoRenew $autoRenew)
     {
-        $input = $request->all();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_AUTORENEW_UPDATE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'auto_renew_min_amt' => 'nullable|string',
-            'auto_renew_amt' => 'nullable|string|max:50',
-            'auto_renewal' => 'nullable|string|max:100',
-        ]);
-
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
+            $validator = Validator::make($input, [
+                'auto_renew_min_amt' => 'nullable|string',
+                'auto_renew_amt' => 'nullable|string|max:50',
+                'auto_renewal' => 'nullable|string|max:100',
             ]);
-        }
 
-        try {
-            $autoRenew->update($input);
-            return response()->json($autoRenew);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
             }
-            else{
-                $message = "Key update error";
+
+            try {
+                $autoRenew->update($input);
+                return response()->json($autoRenew);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "Key update error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
             }
+        }
+        else{
             return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
             ]);
         }
     }
@@ -125,7 +161,16 @@ class AutoRenewController extends Controller
      */
     public function destroy(AutoRenew $autoRenew)
     {
-        $autoRenew->delete();
-        return response()->json();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_AUTORENEW_DESTROY)) {
+            $autoRenew->delete();
+            return response()->json();
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
     }
 }
